==> 01-initial_data_clean.ipynb <==
01-initial_data_clean.ipynb

first run download_convert.sh to download sql file and convert to csv
- Expand predicates with OR operations into individual predicates
- Convert cuis that are entrez ids into cuis
- Change neg props to the same prop with a negative flag
- Make a separate nodes table

import os
import pickle
%matplotlib inline
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
from tqdm import tqdm_notebook
from collections import defaultdict
import itertools
import numpy as np

## This takes too much ram ##
sem_df = pd.read_csv('semmedVER31_R_PREDICATION_to12312017.csv', 
                     usecols=['SUBJECT_CUI', 'PREDICATE','OBJECT_CUI'],
                     #nrows=int(93876633/10),
                     dtype=str
                    )
print(sem_df.shape)
# Group rows together with the same subj -> pred -> obj and keep count of number of pmids
sem_df = sem_df.groupby(['SUBJECT_CUI', 'PREDICATE','OBJECT_CUI']).size().reset_index(name='counts')

# read in csv, group rows together with the same subj -> pred -> obj and keep count of number of pmids
# done in chunks
cols = ['SUBJECT_CUI','PREDICATE','OBJECT_CUI','PMID']
gb_cols = ['SUBJECT_CUI','PREDICATE','OBJECT_CUI']
sem_df = pd.DataFrame(columns=cols)
df_iter = pd.read_csv('semmedVER31_R_PREDICATION_to12312017.csv', dtype=str, usecols=cols, chunksize=10000000)
for chunk in tqdm_notebook(df_iter, total=10):
    chunk.PMID = chunk.PMID.astype("str")
    c = chunk.groupby(gb_cols).PMID.agg(";".join).reset_index()
    sem_df = sem_df.append(c)

# group everything together again
print(len(sem_df))
sem_df = sem_df.groupby(gb_cols).PMID.agg(";".join).reset_index()
print(len(sem_df))

print(sem_df.PMID.str.count(";").value_counts()[:5])

sem_df.head()

sem_df.to_csv("edges1.csv")

#### in semmedDB some subjects and objects of extracted statments contained the pipe character | as an indicator of multiple concepts in the sentence.

# separate out lines with pipes from the rest
multi_start = sem_df['SUBJECT_CUI'].str.contains('|', regex=False)
multi_end = sem_df['OBJECT_CUI'].str.contains('|', regex=False)
pipe_lines = sem_df[multi_start | multi_end].copy()
sem_df = sem_df[~multi_start & ~multi_end]
print('Rows with multiple subjects or objects {:,}'.format(len(pipe_lines)))
print('Rows with only 1 subject AND only 1 object {:,}'.format(len(sem_df)))

pipe_lines.SUBJECT_CUI = pipe_lines.SUBJECT_CUI.str.split('|')
pipe_lines.OBJECT_CUI = pipe_lines.OBJECT_CUI.str.split('|')
pipe_lines.head()

# do the combinations
lines = []
for row in tqdm_notebook(pipe_lines.itertuples(), total=len(pipe_lines)):
    #print(row)
    a = [row.SUBJECT_CUI, row.OBJECT_CUI]
    c = list(itertools.product(*a))
    lines.extend([{'SUBJECT_CUI':x[0], 'PREDICATE':row.PREDICATE, 'OBJECT_CUI':x[1], 'PMID': row.PMID} for x in c])
expanded_df = pd.DataFrame(lines)

print(len(expanded_df))
expanded_df.head()

# append the expanded rows to the original df
print(len(sem_df))
sem_df = sem_df.append(expanded_df, sort=True)
print(len(sem_df))

# Group rows together again and aggregate pmids
sem_df = sem_df.groupby(gb_cols).PMID.agg(";".join).reset_index()
print(len(sem_df))

### There are many CUIs that are not actually cuis. Mike has determined they are (usually/always?) entrez IDs

# are going to get them out of this MRSAT file (download from umls metathersourosuous)
# I did: cat MRSAT.RRF.a* | gzip > MRSAT.RRF.gz
names = list("abcdefghijklmn")
iter_csv = pd.read_csv("MRSAT.RRF.gz", delimiter="|", names=names, index_col=None, chunksize=1000000)
chunks = []
umls_entrez = dict()
for chunk in tqdm_notebook(iter_csv, total=67668372/1000000):
    chunk.fillna(method='ffill', inplace=True)
    chunk = chunk[chunk.i == "ENTREZGENE_ID"]
    d = dict(zip(chunk.a, chunk.k))
    umls_entrez.update(d)

entrez_umls = {v:k for k,v in umls_entrez.items()}
entrez_umls['9987']

sem_df.SUBJECT_CUI = sem_df.SUBJECT_CUI.map(lambda x:entrez_umls[x] if x in entrez_umls else x)
sem_df.OBJECT_CUI = sem_df.OBJECT_CUI.map(lambda x:entrez_umls[x] if x in entrez_umls else x)

noncdf = sem_df[~sem_df.SUBJECT_CUI.str.startswith("C")]
print(len(noncdf))

# dump everything that doesn't starts with a "C"
print(len(sem_df))
sem_df = sem_df[sem_df.SUBJECT_CUI.str.startswith("C")]
sem_df = sem_df[sem_df.OBJECT_CUI.str.startswith("C")]
print(len(sem_df))

# sem_df.to_csv("edges2.csv")
# sem_df = pd.read_csv("edges2.csv", index_col=0)

# change the neg to the same prop without neg and add a neg column# chang 
idx = sem_df["PREDICATE"].str.startswith("NEG_")
sem_df['NEG'] = False
sem_df.loc[idx, 'NEG'] = True
sem_df.loc[idx, 'PREDICATE'] = sem_df[idx].PREDICATE.str.replace("NEG_", "")
sem_df[sem_df.NEG].head()

sem_df.to_csv("edges3.csv")

### Make a nodes table
conso = pd.read_csv("MRCONSO_ENG.RRF.gz", delimiter="|", index_col=None, names = list("abcdefghijklmnopqrs"))
conso = conso[(conso['c'] == "P") & (conso['e'] == "PF")]
conso.head()

node_label = dict(zip(conso.a, conso.o))
print(len(node_label))

nodes = set(sem_df.SUBJECT_CUI) | set(sem_df.OBJECT_CUI)
print(len(nodes))

nodes = pd.DataFrame({"ID":x, "LABEL": node_label.get(x)} for x in nodes)
nodes = nodes.dropna()
print(len(nodes))
nodes.head()

nodes.to_csv("nodes1.csv")
