==> 02-filter_edges.ipynb <==
# remove edges with low counts/occurence, or no good mapping, or not useful

import os
import pickle
%pylab
%matplotlib inline
import pandas as pd
import seaborn as sns
import re
from collections import defaultdict
from tqdm import tqdm

edges = pd.read_csv('edges_sanitized.csv')

nodes = pd.read_csv("nodes_fix_types.csv", index_col=0)
nodes.head()

abv = pd.read_csv("abv.csv")
name_abv = dict(zip(abv.full_name, abv.abv))
abv_name = dict(zip(abv.abv, abv.full_name))

edges.head()

vc = edges.PRED.value_counts()
vc.index = vc.index.map(lambda x:abv_name.get(x))
vc

def drop_edges(df, drop_edges):
    idx = df.query('PRED in @drop_edges').index
    df.drop(idx, inplace=True)

# remove edges we don't need
print(len(edges))
remove_edges = ['compared_with', 'higher_than', 'lower_than', 'different_from', 'different_than', 'same_as',
               'OCCURS_IN', 'PROCESS_OF', 'DIAGNOSES', 'METHOD_OF', 'USES',
               'AUGMENTS', 'ADMINISTERED_TO', 'COMPLICATES']
drop_edges(edges, list(map(name_abv.get, remove_edges)))
print(len(edges))

edge_type_counts = edges.TYPE.value_counts()
print(edge_type_counts.head())
print(edge_type_counts.tail())

# somewhat arbitrarily remove type->pred->type edges with less than 1000 instances
print(len(edge_type_counts))
idx = edge_type_counts[edge_type_counts>1000]
print("keeping {} edges out of {}".format(len(idx), len(edge_type_counts)))
orig_len = len(edges)
edges = edges[edges.TYPE.isin(idx.index)]
print("keeping {} out of {} rows".format(len(edges), orig_len))

edge_type_counts = edges.TYPE.value_counts()

edge_type_counts[edge_type_counts.index.str.contains("cw")]

print(len(nodes))

# remove nodes with no edges
nodes = nodes[nodes.ID.isin(set(list(edges.START_ID) + list(edges.END_ID)))]

print(len(nodes))

vc = edges.PRED.value_counts()
vc.index = vc.index.map(lambda x:abv_name.get(x))
vc

edges.to_csv('edges_filtered.csv', index=False)
nodes.to_csv("nodes_filtered.csv", index=False)
