==> 02-normalize_node_types_to_biolink.ipynb <==
02-normalize_node_types_to_biolink
- For each node, get the UMLS semantic type for each umls cui
(get types from the UMLS mapping file: MRSTY.RRF)
- Map umls semantic types to biolink node types (blm_to_umls_nodes.json)
- Nodes with no matching type are removed
and also special mappings for 
{'protein', 'chemical_substance'} = {"biological_entity"}
{'cell', 'chemical_substance'} = {"biological_entity"}
{'genomic_entity', 'chemical_substance'} = {"biological_entity"}

import os
import pickle
%matplotlib inline
import pandas as pd
import seaborn as sns
import shelve
from itertools import chain
import re
from collections import defaultdict, Counter
from tqdm import tqdm
import requests
from IPython.core.display import display, HTML
display(HTML("<style>.container { width:90% !important; }</style>"))
pd.options.display.max_colwidth = 80

groups = pd.read_csv("https://metamap.nlm.nih.gov/Docs/SemGroups_2013.txt", sep="|",
                    names=['abv', 'group', 'id', 'label'])
groups.head()

print(sorted(list(groups.query("group == 'Disorders'").label)))

names = pd.read_csv("https://metamap.nlm.nih.gov/Docs/SemanticTypes_2013AA.txt", sep="|", 
                   names=['abv', 'ID', 'label'])
names.head()

type_label = dict(zip(names.ID, names.label))

df = pd.read_csv("MRSTY.RRF.gz", sep="|", names=['ID', 'TYPE', 'who', 'the', 'fuck', 'knows'], index_col=False, usecols=['ID', 'TYPE'], dtype=str)
df.head()

id_type = df.groupby("ID").TYPE.aggregate(set).to_dict()

id_type_label = {k:{type_label.get(x) for x in v} for k,v in id_type.items()}
id_type_label['C0000005']

nodes = pd.read_csv("nodes1.csv", index_col=0)
nodes.head()

nodes['umls_type'] = nodes.ID.map(lambda x: id_type.get(x))
nodes['umls_type_label'] = nodes.umls_type.map(lambda x:{type_label.get(y) for y in x} if x else None)

nodes.head()

# toss those with no umls types
print(len(nodes))
nodes.dropna(subset=['umls_type'], inplace=True)
print(len(nodes))

import json

blm_to_umls = json.load(open("blm_to_umls_nodes.json"))
blm_to_umls = {k:set(v) for k,v in blm_to_umls.items()}
blm_to_umls

umls_to_blm_check = defaultdict(set)
umls_to_blm = dict()
for k,vv in blm_to_umls.items():
    for v in vv:
        umls_to_blm_check[v.lower()].add(k.lower())
        umls_to_blm[v.lower()] = k.lower()
assert set(len(x) for x in umls_to_blm_check.values()) == {1}
# umls_to_blm

nodes['blm_type'] = nodes.umls_type_label.map(lambda x: {umls_to_blm.get(xx.lower()) for xx in x})
nodes.blm_type = nodes.blm_type.map(lambda v: {x for x in v if x})
nodes.blm_type = nodes.blm_type.map(lambda v: v if v else pd.np.NAN)
nodes.dropna(subset=['blm_type'], inplace=True)
nodes.head()

nodes[nodes.ID.isin({"C0079904", "C0879593", "C0006875"})]

# the only nodes with two types:
nodes[nodes.blm_type.map(len)>1].blm_type.map(frozenset).value_counts()

# within the nodes with more than 1 blm types:
# what are the most common umls types
nodes2 = nodes[nodes.blm_type.map(len)>1]
Counter(nodes2[nodes2.umls_type_label.map(len)>1].umls_type_label.map(frozenset)).most_common(10)

nodes.loc[nodes.blm_type == {'protein', 'chemical_substance'}, "blm_type"] = {"biological_entity"}
nodes.loc[nodes.blm_type == {'cell', 'chemical_substance'}, "blm_type"] = {"biological_entity"}
nodes.loc[nodes.blm_type == {'genomic_entity', 'chemical_substance'}, "blm_type"] = {"biological_entity"}
nodes.blm_type = nodes.blm_type.map(lambda x:list(x)[0])

nodes[nodes.ID.isin({"C0079904", "C0879593", "C0006875"})]

nodes[nodes.blm_type=="disease_or_phenotypic_feature"].umls_type_label.map(frozenset).value_counts()

nodes.blm_type.value_counts()

nodes.umls_type_label = nodes.umls_type_label.apply("|".join)
nodes.umls_type = nodes.umls_type.apply("|".join)

nodes[nodes.ID.isin({"C0079904", "C0879593", "C0006875", "C0815111"})]

nodes.to_csv("nodes_blm.csv")
