==> 06-neo4j.ipynb <==
06-neo4j
 - reformat for neo4j import

# make the file headers correct
# https://neo4j.com/docs/operations-manual/current/tools/import/file-header-format/

import os
import pickle
%matplotlib inline
import pandas as pd
import seaborn as sns
import shelve
import re
from collections import defaultdict, Counter
from tqdm import tqdm
import requests
from pyquery import PyQuery as pq

# https://biolink.github.io/biolink-model/docs/NamedThing.html
nodes = pd.read_csv("nodes_xref.csv")
nodes.head()

nodes.blm_type.value_counts()

nodes.ID = "UMLS:" + nodes.ID
nodes['category:STRING'] = nodes.blm_type
nodes['id:STRING'] = nodes.ID
nodes.umls_type_label = nodes.umls_type_label.str.replace("|", ";")
nodes.umls_type = nodes.umls_type.str.replace("|", ";")

nodes.rename(columns = {'ID': ':ID', 
                        'LABEL': 'name:STRING', 
                        'blm_type': ':LABEL',
                        'umls_type_label': 'umls_type_label:STRING[]',
                        'umls_type': 'umls_type:STRING[]',
                        'xrefs': 'xrefs:STRING[]'}, inplace=True)

nodes.head()

nodes.to_csv("nodes_neo4j.csv", index=False)

###### EDGES

# https://biolink.github.io/biolink-model/docs/Association.html
edges = pd.read_csv('edges_biolink.csv')

edges['START_ID'] = "UMLS:" + edges.SUBJECT_CUI
edges['END_ID'] = "UMLS:" + edges.OBJECT_CUI

edges['n_pmids'] = edges.PMID.str.count(";")+1

edges.head()

edges.bl_pred = edges.bl_pred.str.lower()
edges.rename(columns = {'START_ID': ':START_ID', 'END_ID': ':END_ID', 
                        'bl_pred': ':TYPE', 'NEG': 'negated',
                       'PMID': 'pmids'}, inplace=True)

edges['is_defined_by'] = "semmeddb"
edges['relation'] = "semmeddb:" + edges[":TYPE"].str.lower()
edges['provided_by'] = "semmeddb_sulab"

del edges['SUBJECT_CUI']
del edges['OBJECT_CUI']

edges.head()

edges.to_csv("edges_neo4j.csv", index=False)

# notes to copy and set up neo4j

scp -i ~/.ssh/Neo4j-AWSMarketplace-Key.pem nodes_neo4j.csv edges_neo4j.csv ubuntu@34.229.55.225:

ssh -i ~/.ssh/Neo4j-AWSMarketplace-Key.pem ubuntu@34.229.55.225
# on aws
sudo su
neo4j stop
rm -rf /var/lib/neo4j/data/databases/graph.db
cd /var/lib/neo4j/import/
neo4j-admin import --nodes nodes_neo4j.csv --relationships edges_neo4j.csv
neo4j start
